Completed Above Average

+++ cpp_files/PracticeSet#1/main_4.cpp
@@ -2,34 +2,35 @@
 #include <cmath>
 #include <iomanip>
 
+// Need to use doubles
+
 struct Vector2D
 {
   Vector2D()
   : x(0),y(0){}
 
-  Vector2D(float x, float y)
+  Vector2D(double x, double y)
   : x(x),y(y){}
 
-  float x, y;
+  double x, y;
 
-  float length()
+  double length()
   {
       return sqrt(x*x + y * y);
   }
 
-  float dotProduct(const Vector2D& other)
+  double dotProduct(const Vector2D& other)
   {
-    float dot = x*other.x + y* other.y;
+    double dot = x*other.x + y* other.y;
     return dot;
   }
-      
 
   Vector2D normalize()
   {
     Vector2D norm_vec = Vector2D(x/length(),y/length());
     return norm_vec;
   }
-   
+
 };
 
 int main()
@@ -40,17 +41,14 @@ int main()
   std::cin >> v1.x >> v1.y >> v2.x >> v2.y;
   std::cout.precision(4);
   std::cout << std::fixed;
-  float dot = v1.dotProduct(v2);
+  double dot = v1.dotProduct(v2);
 
   dot < 0 ? angle = "Obtuse" : dot > 0 ? angle = "Acute" : angle = "Perpendicular";  
 
   std::cout<< "Length of v1: "<< v1.length() <<std::endl;
   std::cout<< "Length of v2: " << v2.length()<<std::endl;
   
-  std::cout<< "Their normalized dot product is "<< v1.normalize().dotProduct(v2.normalize()) << " and they are " <<angle << std::endl;
+  std::cout<< "Their normalized dot product is "<< v1.normalize().dotProduct(v2.normalize()) << " and they are " << angle << std::endl;
 
  // "Length of v1: 1.000 \n Length of v2: 1.000 \n Their normalized dot product is 0.000 and they are Perpendicular"
-
-
-
 }
